backend/requirements.txt
django==4.0.*
djangorestframework==3.13.*
Pillow==9.0.*
pytest-django==4.5.*